Move OpenGL initialization code to GL46_Renderer
